requirements.txt
accelerate==1.13.0
bitsandbytes==0.46.1
datasets==4.8.4
google-genai==1.62.0
openai==2.14.0
peft==0.18.1
pydantic==2.11.7
python-dotenv==1.1.1
PyYAML==6.0.2
torch==2.8.0
tqdm==4.67.1
transformers==5.5.0
trl==1.0.0
vllm==0.10.2
wandb==0.22.1
